CLeanup duplication processRequest and auth/App.vue and export findBestWildcardMatch
