+++ src/entrypoints/auth/App.vue
@@ -8,7 +8,7 @@ import { openOptions } from '@/utils/extension.ts'
 import { getSession, saveKeyValue } from '@/utils/options.ts'
 import { useOptions } from '@/composables/useOptions.ts'
 import { showToast } from '@/composables/useToast.ts'
-import { Hosts, matchesWildcard, countWildcards } from '@/utils/hosts.ts'
+import { Hosts, findBestWildcardMatch } from '@/utils/hosts.ts'
 import ToastAlerts from '@/components/ToastAlerts.vue'
 import BackToTop from '@/components/BackToTop.vue'
 import OptionsOffscreen from '@/components/OptionsOffscreen.vue'
@@ -93,6 +93,14 @@ function saveCredsChange(event?: Event) {
   sessionStorage.setItem(hostRef.value, saveCreds.value ? '1' : '0')
 }
 
+async function populateFields(credsStr: string) {
+  const [username, password] = parseCreds(credsStr)
+  userRef.value = username
+  await nextTick()
+  usernameEl.value?.select()
+  passRef.value = password
+}
+
 onMounted(async () => {
   // NOTE: Copied from VanillaJS...
   const searchParams = new URLSearchParams(window.location.search)
@@ -117,44 +125,19 @@ onMounted(async () => {
   debug('session:', session)
 
   if (creds) {
-    debug('if creds:', creds)
     hasSavedCreds.value = true
     if (creds !== 'ignored') {
-      const [username, password] = parseCreds(creds)
-      userRef.value = username
-      debug('usernameEl.value:', usernameEl.value)
-      await nextTick()
-      usernameEl.value?.select()
-      passRef.value = password
+      debug('if creds:', creds)
+      await populateFields(creds)
     }
   } else if (hostRef.value in session) {
     debug('else hostRef.value in session:', hostRef.value)
-    const [username, password] = parseCreds(session[hostRef.value])
-    userRef.value = username
-    debug('usernameEl.value:', usernameEl.value)
-    await nextTick()
-    usernameEl.value?.select()
-    passRef.value = password
+    await populateFields(session[hostRef.value])
   } else {
-    let bestMatch: string | undefined
-    let bestSpecificity = Infinity
-    for (const [pattern, creds] of Object.entries(session)) {
-      if (!pattern.includes('*')) continue
-      if (matchesWildcard(hostRef.value, pattern)) {
-        const specificity = countWildcards(pattern)
-        if (specificity < bestSpecificity) {
-          bestSpecificity = specificity
-          bestMatch = creds
-        }
-      }
-    }
+    const bestMatch = findBestWildcardMatch(hostRef.value, session)
     if (bestMatch) {
       debug('session wildcard match:', bestMatch)
-      const [username, password] = parseCreds(bestMatch)
-      userRef.value = username
-      await nextTick()
-      usernameEl.value?.select()
-      passRef.value = password
+      await populateFields(bestMatch)
     }
   }
