+++ src/entrypoints/content/index.ts
@@ -1,6 +1,6 @@
 import { i18n } from '#imports'
 import { defineContentScript } from 'wxt/utils/define-content-script'
-import { Hosts, matchesWildcard, countWildcards } from '@/utils/hosts.ts'
+import { Hosts, findBestWildcardMatch } from '@/utils/hosts.ts'
 
 // TODO: Logging
 
@@ -43,28 +43,6 @@ async function onChanged(changes: Record<string, any>) {
     const newWildcard = findBestWildcardMatch(url.host, wildcardItems.newValue)
     if (oldWildcard !== newWildcard) await processCreds(newWildcard)
   }
-}
-
-function findBestWildcardMatch(
-  host: string,
-  patterns: Record<string, string> | undefined,
-): string | undefined {
-  if (!patterns) return undefined
-  let bestMatch: string | undefined
-  let bestSpecificity = Infinity
-
-  for (const [pattern, creds] of Object.entries(patterns)) {
-    if (!pattern.includes('*')) continue
-    if (matchesWildcard(host, pattern)) {
-      const specificity = countWildcards(pattern)
-      if (specificity < bestSpecificity) {
-        bestSpecificity = specificity
-        bestMatch = creds
-      }
-    }
-  }
-
-  return bestMatch
 }
 
 async function processCreds(creds: any) {
